pyproject.toml
[project]
name = "python-project-83"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12.2"
dependencies = [
    "dotenv>=0.9.9",
    "flask>=3.1.2",
    "gunicorn>=23.0.0",
    "psycopg2-binary>=2.9.11",
    "requests>=2.32.5",
    "validators>=0.35.0",
    "bs4 (>=0.0.2,<0.0.3)",
    "lxml (>=6.0.2,<7.0.0)",
]

[dependency-groups]
dev = [
    "ipython>=9.8.0",
    "ruff>=0.14.8",
]
